refactor : chatting 단어를 chatroom으로 모두 교체 #59
