+++ src/components/ChatroomDetailContents/index.tsx
@@ -5,7 +5,7 @@ import { v4 as createRandomKey } from 'uuid';
 import Chat from '@components/Chat';
 import * as S from '@components/ChatroomDetailContents/ChatroomDetailContents.style';
 import { TestChatroomDetailChatsType } from '@pages/ChatroomDetail/chatroomDetailData';
-import { chattingConnect } from '@pages/ChatroomDetail/socket';
+import { chatroomConnect } from '@pages/ChatroomDetail/socket';
 
 type ChatroomDetailContentsPropsType = {
   chats: TestChatroomDetailChatsType;
@@ -14,16 +14,16 @@ type ChatroomDetailContentsPropsType = {
 const ChatroomDetailContents = ({ chats }: ChatroomDetailContentsPropsType) => {
   // const [date, setDate] = useState('');
   const [curChats, setCurChats] = useState(chats);
-  const chattingLogs = curChats.map((info) => <Chat {...info} key={createRandomKey()} />);
+  const chatroomLogs = curChats.map((info) => <Chat {...info} key={createRandomKey()} />);
 
   useMemo(() => {
-    chattingConnect({ setter: setCurChats });
+    chatroomConnect({ setter: setCurChats });
   }, []);
 
   return (
     <S.Wrapper>
       <S.Date>2021년 5월 29일</S.Date>
-      <S.Chats>{chattingLogs}</S.Chats>
+      <S.Chats>{chatroomLogs}</S.Chats>
     </S.Wrapper>
   );
 };
