+++ src/components/ChatroomDetailHeader/index.tsx
@@ -3,7 +3,7 @@ import { useNavigate } from 'react-router-dom';
 import * as S from '@components/ChatroomDetailHeader/ChatroomDetailHeader.style';
 import Icon from '@components/common/Icon';
 
-const ChattingDatailHeader = () => {
+const ChatroomDatailHeader = () => {
   const navigate = useNavigate();
 
   return (
@@ -18,4 +18,4 @@ const ChattingDatailHeader = () => {
   );
 };
 
-export default ChattingDatailHeader;
+export default ChatroomDatailHeader;
