+++ src/components/ChatroomError/index.tsx
@@ -1,0 +1,5 @@
+const ChatroomError = ({ mention }: { mention: string }) => {
+  return <div>{mention}</div>;
+};
+
+export default ChatroomError;
