--- src/components/ChattingError/index.tsx
@@ -1,5 +1,0 @@
-const ChattingError = ({ mention }: { mention: string }) => {
-  return <div>{mention}</div>;
-};
-
-export default ChattingError;
